tools/requirements.txt
psutil>=5.9.0
pyserial>=3.5
requests>=2.31.0
